Ported the density_calculator class to C. One test
Only test has been implemented using the platformio unit testing
engine. Unfortunately, the unit testing engine requires a subscription
for Platformio plus, which I only have a 30 day trial for. Also, the
testing doesn't seem to be as accurate as I would like. I will attempt
to add googletest as a testing framework instead, but platformio plus
will still be required to run the unit tests on the device itself, which
may be worth the cost. In any case, more tests are required for
density_calculator and especially tests for interpolation since that
will be used extensively for the rest of the classes