Feature complete code for launch on February 1st
There are two new files in this commit for the acceleration calculator,
a header file and a test file. The acceleration calculator has a header
file only since it relies on a template for its array sizes. This may
have to change in later. There also is a lot of change in the testing
files so that the teensy and native platforms can run the same code.
I also updated the platformio.ini file to reflect that we're using the
teensy3.5 and not the teensy3.2.