Finished initial implementation of Verlet Integrator
The Verlet Integrator is simplified from the python version since the
use case has narrowed in scope. We're no longer running the code while
the engine is firing so there's no need to keep track of engine data
or any of that stuff. This initial implementation should serve well
enough for the use case of the rocket.